pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "binance-scalper-bot"
version = "0.1.0"
description = "High-frequency trading bot for Binance"
authors = [
    {name = "Developer"}
]
requires-python = ">=3.12"
dependencies = [
    "ccxt (>=4.5.37,<5.0.0)",
    "pandas (>=2.0,<3.0)",
    "pandas-ta (>=0.4.71b0,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyarrow (>=23.0.0,<24.0.0)",
    "fastparquet (>=2025.12.0,<2026.0.0)",
    "websocket-client (>=1.9.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
